chore: refactor <SigupPage/>
