+++ frontend/app/(auth)/signup/page.tsx
@@ -1,48 +1,16 @@
 /* eslint-disable */
 "use client";
 import Link from "next/link";
-import { useState } from "react";
 
 import Button from "@/lib/components/ui/Button";
 import Card from "@/lib/components/ui/Card";
 import Field from "@/lib/components/ui/Field";
 import PageHeading from "@/lib/components/ui/PageHeading";
-import { useSupabase } from "@/lib/context/SupabaseProvider";
-import { useToast } from "@/lib/hooks/useToast";
-import { useEventTracking } from "@/services/analytics/useEventTracking";
+import { useSignUp } from "./hooks/useSignUp";
 
 export default function SignUp() {
-  const { supabase } = useSupabase();
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [isPending, setIsPending] = useState(false);
-  const { track } = useEventTracking();
-
-
-  const { publish } = useToast();
-  const handleSignUp = async () => {
-    void track("SIGNUP")
-    setIsPending(true);
-    const { data, error } = await supabase.auth.signUp({
-      email: email,
-      password: password,
-    });
-
-    if (error) {
-      console.error("Error signing up:", error.message);
-      publish({
-        variant: "danger",
-        text: `Error signing up: ${error.message}`,
-      });
-    } else if (data) {
-      publish({
-        variant: "success",
-        text: "Confirmation Email sent, please check your email",
-      });
-    }
-    setIsPending(false);
-  };
-
+  const { handleSignUp, isPending, email, password, setEmail, setPassword } =
+    useSignUp();
   return (
     <main>
       <section className="min-h-[80vh] w-full h-full outline-none flex flex-col gap-5 items-center justify-center p-6">
@@ -60,6 +28,7 @@ export default function SignUp() {
               required
               type="email"
               placeholder="Email"
+              value={email}
               onChange={(e) => setEmail(e.target.value)}
             />
             <Field
